+++ tests/src/parse_string_tests.cpp
@@ -35,12 +35,11 @@ private:
 TEST_CASE("test_parse_small_string1")
 {
     std::string input = "\"String\"";
-    std::istringstream is(input);
 
     json_decoder<json> decoder;
     try
     {
-        json_reader reader(is,decoder);
+        json_reader reader(input,decoder);
         reader.read_next();
     }
     catch (const std::exception&)
@@ -52,12 +51,11 @@ TEST_CASE("test_parse_small_string1")
 TEST_CASE("test_parse_small_string2")
 {
     std::string input = "\"Str\\\"ing\"";
-    std::istringstream is(input);
 
     json_decoder<json> decoder;
     try
     {
-        json_reader reader(is, decoder);
+        json_reader reader(input, decoder);
         reader.read_next();
     }
     catch (const std::exception&)