+++ ChangeLog.md
@@ -3,6 +3,7 @@ master
 
 Name changes (non breaking)
 
+- The name `json_stream_reader` has been deprecated and replaced with `json_staj_reader`.
 - The name `stream_event_type` has been deprecated and replaced with `staj_event_type`
 - The names `basic_stream_event` (`stream_event`) have been deprecated and replaced with `basic_staj_event` (`staj_event`)
 - The names `basic_stream_filter` (`stream_filter`) have been deprecated and replaced with `basic_staj_filter` (`staj_filter`)